+++ cli/tests/test_train_pretraining_file_flags.py
@@ -27,7 +27,6 @@ TRAIN_SCRIPT_CASES = [
     "train_llama_fast",
     "train_llama_megakernel",
     "train_gpt2",
-    "train_nanogpt",
     "train_semantic_router_moe",
 ]